move player lose event logic to player manager, away from next turn validator
