+++ Assets/Scripts/NextTurnValidator.cs
@@ -11,35 +11,12 @@ public class NextTurnValidator : MonoBehaviour
 {
     [field: SerializeField]
     public List<PlayerResource> AllResources { get; set; }
-    private GameManager GameManager;
-    [SerializeField] private int maxC02 = 500;
-    [SerializeField] private int maxTemperature = 50;
-
-    [SerializeField] private GameEvent _PlayerLoseEvent;
-
     public GameEvent NextTurnClickedEvent;
-
-    void Start()
-    {
-        GameManager = FindAnyObjectByType<GameManager>();
-    }
 
     public void OnValidatingNextTurn(Object _)
     {
         if (!HasEnoughResources())
         {
-            return;
-        }
-
-        if (GameManager.GetCO2() > maxC02)
-        {
-            _PlayerLoseEvent.Raise();
-            return;
-        }
-
-        if (GameManager.GetTemperature() > maxTemperature)
-        {
-            _PlayerLoseEvent.Raise();
             return;
         }
